Jsonifier Release v0.9.7
-Removed a bunch of superfluous "inline" calls.
-Updated a few types.
-Updated the Vector, StringBase and StringViewBase classes to fill their implementations out.
-Updated the Iterator and ConstIterator implementations as well as the iterator-related members of the String, StringView and Vector classes.

+++ Include/jsonifier/RawVector.hpp
@@ -1,22 +1,22 @@
 /*
-		MIT License
+	MIT License
 
 	Copyright (c) 2023 RealTimeChris
 
-	Permission is hereby granted, free of charge, to any person obtaining a copy of this 
-	software and associated documentation files (the "Software"), to deal in the Software 
-	without restriction, including without limitation the rights to use, copy, modify, merge, 
-	publish, distribute, sublicense, and/or sell copies of the Software, and to permit 
+	Permission is hereby granted, free of charge, to any person obtaining a copy of this
+	software and associated documentation files (the "Software"), to deal in the Software
+	without restriction, including without limitation the rights to use, copy, modify, merge,
+	publish, distribute, sublicense, and/or sell copies of the Software, and to permit
 	persons to whom the Software is furnished to do so, subject to the following conditions:
 
-	The above copyright notice and this permission notice shall be included in all copies or 
+	The above copyright notice and this permission notice shall be included in all copies or
 	substantial portions of the Software.
 
-	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, 
-	INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR 
-	PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE 
-	FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR 
-	OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
+	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+	INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+	PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
+	FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+	OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 	DEALINGS IN THE SOFTWARE.
 */
 /// https://github.com/RealTimeChris/Jsonifier
@@ -27,23 +27,23 @@
 
 namespace JsonifierInternal {
 
-	template<typename ValueType, size_t N> class RawVector {
+	template<typename ValueType, uint64_t N> class RawVector {
 	  public:
-		using value_type = ValueType;
-		using reference = value_type&;
+		using value_type	  = ValueType;
+		using reference		  = value_type&;
 		using const_reference = const value_type&;
-		using pointer = value_type*;
-		using const_pointer = const value_type*;
-		using iterator = pointer;
-		using const_iterator = const_pointer;
-		using size_type = size_t;
+		using pointer		  = value_type*;
+		using const_pointer	  = const value_type*;
+		using iterator		  = pointer;
+		using const_iterator  = const_pointer;
+		using size_type		  = uint64_t;
 		using difference_type = std::ptrdiff_t;
 
-		inline constexpr RawVector() noexcept = default;
-		inline constexpr RawVector& operator=(RawVector&&) noexcept = default;
-		inline constexpr RawVector(RawVector&&) noexcept = default;
+		inline constexpr RawVector() noexcept							 = default;
+		inline constexpr RawVector& operator=(RawVector&&) noexcept		 = default;
+		inline constexpr RawVector(RawVector&&) noexcept				 = default;
 		inline constexpr RawVector& operator=(const RawVector&) noexcept = default;
-		inline constexpr RawVector(const RawVector&) noexcept = default;
+		inline constexpr RawVector(const RawVector&) noexcept			 = default;
 
 		inline constexpr RawVector(size_type count, const auto& value) : dsize(count) {
 			for (size_type x = 0; x < N; ++x)
@@ -104,7 +104,7 @@ namespace JsonifierInternal {
 
 	  protected:
 		value_type data[N] = {};
-		size_type dsize = 0;
+		size_type dsize	   = 0;
 	};
 
 }
